refactor Monaco file
extract editor actions to dedicated file
refactor types
